refactoring of classifier module
