+++ mriqc/classifier/__init__.py
@@ -52,7 +52,4 @@ indexing columns and the predicted quality label under the ``prediction`` column
     cv/data
 
 """
-from __future__ import print_function
-from __future__ import division
-from __future__ import absolute_import
-from __future__ import unicode_literals
+from __future__ import print_function, division, absolute_import, unicode_literals
